+++ importing/finalizepmx.py
@@ -26,6 +26,7 @@ def rename_and_merge_outfits():
     body.name = 'Body'
     body.modifiers[0].show_in_editmode = True
     body.modifiers[0].show_on_cage = True
+    body.modifiers[0].show_expanded = False
     
     #Deselect all objects
     bpy.ops.object.select_all(action='DESELECT')
@@ -49,6 +50,7 @@ def rename_and_merge_outfits():
             outfit.modifiers[0].object = armature
             outfit.modifiers[0].show_in_editmode = True
             outfit.modifiers[0].show_on_cage = True
+            outfit.modifiers[0].show_expanded = False
             
             idx += 1